--- disabledTests/Api/Admin/PromotionsTest.php
@@ -1,156 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Tests\Api\Admin;
-
-use Sylius\Component\Core\Model\PromotionInterface;
-use Sylius\Tests\Api\JsonApiTestCase;
-use Sylius\Tests\Api\Utils\AdminUserLoginTrait;
-use Symfony\Component\HttpFoundation\Response;
-
-final class PromotionsTest extends JsonApiTestCase
-{
-    use AdminUserLoginTrait;
-
-    /** @test */
-    public function it_gets_promotion_coupons(): void
-    {
-        $fixtures = $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'channel.yaml', 'promotion/promotion.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        /** @var PromotionInterface $promotion */
-        $promotion = $fixtures['promotion_1_off'];
-
-        $this->client->request(
-            method: 'GET',
-            uri: sprintf('/api/v2/admin/promotions/%s/coupons', $promotion->getCode()),
-            server: $header,
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/get_promotion_coupons_response',
-            Response::HTTP_OK,
-        );
-    }
-
-    /** @test */
-    public function it_updates_promotion(): void
-    {
-        $fixtures = $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'channel.yaml', 'promotion/promotion.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        /** @var PromotionInterface $promotion */
-        $promotion = $fixtures['promotion_50_off'];
-
-        $this->client->request(
-            method: 'PUT',
-            uri: sprintf('/api/v2/admin/promotions/%s', $promotion->getCode()),
-            server: $header,
-            content: json_encode([
-                'name' => 'Christmas',
-                'code' => 'new_code',
-                'appliesToDiscounted' => true,
-                'exclusive' => true,
-                'usageLimit' => 11,
-                'couponBased' => false,
-                'rules' => [
-                    [
-                        'type' => CartQuantityRuleChecker::TYPE,
-                        'configuration' => [
-                            'count' => 1,
-                        ],
-                    ],
-                ],
-                'actions' => [
-                    [
-                        'type' => FixedDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'amount' => 2,
-                            ],
-                            'MOBILE' => [
-                                'amount' => 2,
-                            ],
-                        ],
-                    ],
-                ],
-                'channels' => [
-                    '/api/v2/admin/channels/MOBILE',
-                ],
-                'translations' => ['en_US' => [
-                    '@id' => sprintf('/api/v2/admin/promotion-translations/%s', $promotion->getTranslation('en_US')->getId()),
-                    'label' => 'Christmas',
-                ]],
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/put_promotion_response',
-            Response::HTTP_OK,
-        );
-    }
-
-    /** @test */
-    public function it_updates_promotion_to_last_priority_when_priority_is_minus_one(): void
-    {
-        $fixtures = $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'channel.yaml', 'promotion/promotion.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        /** @var PromotionInterface $promotion */
-        $promotion = $fixtures['promotion_50_off'];
-
-        $this->client->request(
-            method: 'PUT',
-            uri: sprintf('/api/v2/admin/promotions/%s', $promotion->getCode()),
-            server: $header,
-            content: json_encode([
-                'priority' => -1,
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/put_promotion_to_last_priority_when_priority_is_minus_one_response',
-            Response::HTTP_OK,
-        );
-    }
-
-    /** @test */
-    public function it_does_not_update_a_promotion_with_duplicate_locale_translation(): void
-    {
-        $fixtures = $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'channel.yaml', 'promotion/promotion.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        /** @var PromotionInterface $promotion */
-        $promotion = $fixtures['promotion_50_off'];
-
-        $this->client->request(
-            method: 'PUT',
-            uri: sprintf('/api/v2/admin/promotions/%s', $promotion->getCode()),
-            server: $header,
-            content: json_encode([
-                'translations' => ['en_US' => [
-                    'label' => 'Christmas',
-                ]],
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/put_promotion_with_duplicate_locale_translation',
-            Response::HTTP_UNPROCESSABLE_ENTITY,
-        );
-    }
-}
