+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/filters.xml
@@ -314,14 +314,14 @@
 <!--            <tag name="api_platform.filter" />-->
 <!--        </service>-->
 
-        <service id="sylius_api.promotion_coupon_search_filter" parent="api_platform.doctrine.orm.search_filter" public="true">
+        <service id="sylius_api.search_filter.promotion" parent="api_platform.doctrine.orm.search_filter" public="true">
             <argument type="collection">
                 <argument key="coupons.code">exact</argument>
             </argument>
             <tag name="api_platform.filter" />
         </service>
 
-        <service id="sylius_api.promotion_order_filter" parent="api_platform.doctrine.orm.order_filter" public="true">
+        <service id="sylius_api.order_filter.promotion" parent="api_platform.doctrine.orm.order_filter" public="true">
             <argument type="collection">
                 <argument key="priority" />
             </argument>
@@ -408,7 +408,7 @@
             <tag name="api_platform.filter" />
         </service>
 
-        <service id="sylius_api.exists_filter.archived_at" parent="api_platform.doctrine.orm.exists_filter" public="true">
+        <service id="sylius_api.exists_filter.promotion" parent="api_platform.doctrine.orm.exists_filter" public="true">
             <argument type="collection">
                 <argument key="archivedAt" />
             </argument>
