+++ tests/Api/Admin/PromotionsTest.php
@@ -73,6 +73,28 @@ final class PromotionsTest extends JsonApiTestCase
     }
 
     /** @test */
+    public function it_gets_promotion_coupons(): void
+    {
+        $fixtures = $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'channel.yaml', 'promotion/promotion.yaml']);
+        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
+
+        /** @var PromotionInterface $promotion */
+        $promotion = $fixtures['promotion_1_off'];
+
+        $this->client->request(
+            method: 'GET',
+            uri: sprintf('/api/v2/admin/promotions/%s/coupons', $promotion->getCode()),
+            server: $header,
+        );
+
+        $this->assertResponse(
+            $this->client->getResponse(),
+            'admin/promotion/get_promotion_coupons_response',
+            Response::HTTP_OK,
+        );
+    }
+
+    /** @test */
     public function it_creates_promotion(): void
     {
         $this->loadFixturesFromFiles([
